+++ web-views/globals.d.ts
@@ -1,5 +1,7 @@
 export declare global {
 	const tsVscode: {
 		postMessage: (msg: { type: string; value: string }) => void;
+		setState: (state: any) => void;
+		getState: () => undefined | any;
 	};
 }