+++ CHANGELOG.md
@@ -3,3 +3,10 @@
 ## Version 0.0.1
 
 - Initial release
+
+## Version 0.1.0
+
+- Bug fixes
+- Added the ability to use STOMPjs without the SockJs-Client
+- WebView state now will persist after closing the panel
+  > **New Connection Is Made** the old one is lost